if salaru is float it is hard to check

+++ schemas/salary_schema.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel, validator, constr
 
 
 class Salary(BaseModel):
-    salary: float
+    salary: int
     salary_rise_date: constr(regex=r'^[0-9][0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]$') = '2023-09-22'
